+++ lib/screens/home_page/editor_pane/details_card/code_pane.dart
@@ -1,4 +1,3 @@
-import 'package:apidash/utils/convert_utils.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:apidash/providers/providers.dart';
@@ -45,85 +44,6 @@ class _CodePaneState extends ConsumerState<CodePane> {
             .read(codegenLanguageStateProvider.notifier)
             .update((state) => value!);
       },
-    );
-  }
-}
-
-class ViewCodePane extends StatefulWidget {
-  const ViewCodePane({
-    super.key,
-    required this.code,
-    required this.codegenLanguage,
-    required this.onChangedCodegenLanguage,
-  });
-
-  final String code;
-  final CodegenLanguage codegenLanguage;
-  final Function(CodegenLanguage?) onChangedCodegenLanguage;
-
-  @override
-  State<ViewCodePane> createState() => _ViewCodePaneState();
-}
-
-class _ViewCodePaneState extends State<ViewCodePane> {
-  @override
-  Widget build(BuildContext context) {
-    var codeTheme = Theme.of(context).brightness == Brightness.light
-        ? kLightCodeTheme
-        : kDarkCodeTheme;
-    final textContainerdecoration = BoxDecoration(
-      color: Color.alphaBlend(
-          (Theme.of(context).brightness == Brightness.dark
-                  ? Theme.of(context).colorScheme.onPrimaryContainer
-                  : Theme.of(context).colorScheme.primaryContainer)
-              .withOpacity(kForegroundOpacity),
-          Theme.of(context).colorScheme.surface),
-      border: Border.all(color: Theme.of(context).colorScheme.surfaceVariant),
-      borderRadius: kBorderRadius8,
-    );
-
-    return Padding(
-      padding: kP10,
-      child: Column(
-        children: [
-          SizedBox(
-            height: kHeaderHeight,
-            child: Row(
-              children: [
-                Expanded(
-                  child: Text(
-                    "Code",
-                    style: Theme.of(context).textTheme.titleMedium,
-                  ),
-                ),
-                DropdownButtonCodegenLanguage(
-                  codegenLanguage: widget.codegenLanguage,
-                  onChanged: widget.onChangedCodegenLanguage,
-                ),
-                CopyButton(toCopy: widget.code),
-                SaveInDownloadsButton(
-                  content: stringToBytes(widget.code),
-                  mimeType: "application/vnd.dart",
-                )
-              ],
-            ),
-          ),
-          kVSpacer10,
-          Expanded(
-            child: Container(
-              width: double.maxFinite,
-              padding: kP8,
-              decoration: textContainerdecoration,
-              child: CodeGenPreviewer(
-                code: widget.code,
-                theme: codeTheme,
-                language: 'dart',
-                textStyle: kCodeStyle,
-              ),
-            ),
-          ),
-        ],
-      ),
     );
   }
 }
