+++ test/widgets/codegen_previewer_test.dart
@@ -1,7 +1,7 @@
-import 'package:apidash/screens/home_page/editor_pane/details_card/code_pane.dart';
+import 'package:apidash/widgets/widgets.dart'
+    show ViewCodePane, CodeGenPreviewer;
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:apidash/widgets/codegen_previewer.dart';
 import 'package:apidash/consts.dart';
 import '../test_consts.dart';
